examples/qpid-proton/pyhton/tutorial/pyproject.toml
[tool.poetry]
name = "tutorial"
version = "0.1.0"
description = ""
authors = ["mrmodolo <marcelo.rmodolo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-qpid-proton = "^0.33.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
